third_party/WebKit/LayoutTests/http/tests/inspector/bindings/contentscripts-bindings-multiple-frames.html

<head>
<script src="../inspector-test.js"></script>
<script src="./bindings-test.js"></script>

<script>
async function test()
{
    InspectorTest.markStep('initialWorkspace');
    var snapshot = InspectorTest.dumpWorkspace();

    InspectorTest.markStep('attachFrame1');
    await InspectorTest.attachFrame('frame1', './resources/contentscript-frame.html', '_test_attachFrame1.js');
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('attachFrame2');
    await InspectorTest.attachFrame('frame2', './resources/contentscript-frame.html', '_test_attachFrame2.js');
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('detachFrame1');
    await InspectorTest.detachFrame('frame1', '_test_detachFrame1.js');
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('detachFrame2');
    await InspectorTest.detachFrame('frame2', '_test_detachFrame2.js');
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.completeTest();
}

</script>
</head>
<body onload="runTest()">
<p>
Verify that SourceMap bindings are generating UISourceCodes properly.
</p>
</body>
